refactor: split Filter component into Filter and AddFilterButton
